+++ mlruns/0/1d7303102d27468b92c77d471fd88367/inputs/8230adb4d5e1a02e4973560c9d449c8f/meta.yaml
@@ -1,0 +1,6 @@
+destination_id: 201dfd781d08dc93ccb58c708b4ff9af
+destination_type: RUN
+source_id: 201dfd781d08dc93ccb58c708b4ff9af
+source_type: DATASET
+tags:
+  mlflow.data.context: train
