+++ mlruns/0/1d7303102d27468b92c77d471fd88367/meta.yaml
@@ -1,0 +1,15 @@
+artifact_uri: mlflow-artifacts:/0/1d7303102d27468b92c77d471fd88367/artifacts
+end_time: 1734472114548
+entry_point_name: ''
+experiment_id: '0'
+lifecycle_stage: active
+run_id: 1d7303102d27468b92c77d471fd88367
+run_name: crawling-stoat-204
+run_uuid: 1d7303102d27468b92c77d471fd88367
+source_name: ''
+source_type: 4
+source_version: ''
+start_time: 1734472110075
+status: 4
+tags: []
+user_id: lenovo
