+++ mlruns/0/1d7303102d27468b92c77d471fd88367/params/multi_class
@@ -1,0 +1,1 @@
+deprecated
