+++ mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/artifacts/logistic/python_env.yaml
@@ -1,0 +1,7 @@
+python: 3.11.9
+build_dependencies:
+- pip==24.3.1
+- setuptools==65.5.0
+- wheel
+dependencies:
+- -r requirements.txt
