+++ mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/artifacts/logistic/requirements.txt
@@ -1,0 +1,7 @@
+mlflow==2.19.0
+cloudpickle==3.1.0
+numpy==2.2.0
+pandas==2.2.3
+psutil==6.1.0
+scikit-learn==1.6.0
+scipy==1.14.1
