+++ mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/meta.yaml
@@ -1,0 +1,15 @@
+artifact_uri: file:///C:/Users/lenovo/Desktop/Cours/034_MLOPS/mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/artifacts
+end_time: 1734472121975
+entry_point_name: ''
+experiment_id: '585590587743873505'
+lifecycle_stage: active
+run_id: 1ebf76dfa25c49a586f527a0f0c640ae
+run_name: mercurial-sheep-636
+run_uuid: 1ebf76dfa25c49a586f527a0f0c640ae
+source_name: ''
+source_type: 4
+source_version: ''
+start_time: 1734472114694
+status: 3
+tags: []
+user_id: lenovo
