+++ mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/metrics/accuracy
@@ -1,0 +1,1 @@
+1734472121929 0.62 0
