+++ mlruns/585590587743873505/1ebf76dfa25c49a586f527a0f0c640ae/tags/mlflow.log-model.history
@@ -1,0 +1,1 @@
+[{"run_id": "1ebf76dfa25c49a586f527a0f0c640ae", "artifact_path": "logistic", "utc_time_created": "2024-12-17 21:48:34.787738", "model_uuid": "12aebfd457a146c0b00dbab26b88f2e8", "flavors": {"python_function": {"model_path": "model.pkl", "predict_fn": "predict", "loader_module": "mlflow.sklearn", "python_version": "3.11.9", "env": {"conda": "conda.yaml", "virtualenv": "python_env.yaml"}}, "sklearn": {"pickled_model": "model.pkl", "sklearn_version": "1.6.0", "serialization_format": "cloudpickle", "code": null}}}]
