+++ mlruns/585590587743873505/78ef88d5adf849cbb55901c0e12bc4d8/meta.yaml
@@ -1,0 +1,15 @@
+artifact_uri: file:///C:/Users/lenovo/Desktop/Cours/034_MLOPS/mlruns/585590587743873505/78ef88d5adf849cbb55901c0e12bc4d8/artifacts
+end_time: 1734472122008
+entry_point_name: ''
+experiment_id: '585590587743873505'
+lifecycle_stage: active
+run_id: 78ef88d5adf849cbb55901c0e12bc4d8
+run_name: luxuriant-whale-926
+run_uuid: 78ef88d5adf849cbb55901c0e12bc4d8
+source_name: ''
+source_type: 4
+source_version: ''
+start_time: 1734472114607
+status: 3
+tags: []
+user_id: lenovo
