+++ mlruns/585590587743873505/78ef88d5adf849cbb55901c0e12bc4d8/tags/mlflow.source.type
@@ -1,0 +1,1 @@
+LOCAL
